[Fix] Handle args with the default parser + remove unused file
